:hammer: Keep a parallel copy of entry options to work from
The main working of the directory navigation widget is fine, but because I
want to show/hide hidden directories, and because I'll be wanting to allow
for sorting, etc, here I load everything into a parallel list of options and
then display from that list. When it comes to showing/hiding hidden items,
rather than reload everything, I just clear the display and repopulate.

This saves going back to the filesystem.