:lipstick: Don't style text as $text for now
I think there's some confusion going on with $text between light/dark mode,
and the way partial component styles work. This might actually be more of an
issue with how OptionList can support someone inheriting from it and further
styling the prompts. I need to dig more into this.

For now though, *not* setting the style of the text solves the problem.

See #1.