:hammer: Have cancel be a result of None
My thinking here is that with the fspicker dialogs the caller may 100% want
to know that the user performed a cancel, as it might indicate they want to
back up from some operation or mode or something. Taking this approach, if
they don't care, code that does use the result can just "is not None"-guard
things.